+++ discord/ext/commands/core.py
@@ -43,6 +43,7 @@ from typing import (
 )
 
 import discord
+import discord.utils.private
 from ...utils.private import evaluate_annotation
 from discord import utils
 from ...utils import Undefined
@@ -1155,7 +1156,7 @@ class Command(_BaseCommand, Generic[CogT, P, T]):
                 # since we have no checks, then we just return True.
                 return True
 
-            return await discord.utils.async_all(predicate(ctx) for predicate in predicates)  # type: ignore
+            return await discord.utils.private.async_all(predicate(ctx) for predicate in predicates)  # type: ignore
         finally:
             ctx.command = original
